Upgrade OpenZeppelin to the latest 4.x version
The project is on OpenZeppelin 4 and the latest version currently
available is 4.9.6. This change will also allow us to use
Ownable2StepUpgradeable that was not available in OpenZeppelin 4.6. The
upgrade required increasing the version of
@openzeppelin/hardhat-upgrades plugin given the bug in the previously
used 1.20.0 version not resolving the correct version of dependencies
used (OZ contracts vs contracts-upgradeable) and in our case complaining
about the delegatecall used in the OZ's AddressUpgradeable code. The bug
was fixed in 1.20.4 version of the plugin. This also required updates in
the deployment tests as some functions are no longer available in the
upgraded version.

https://forum.openzeppelin.com/t/interacting-with-uups-upgradeable-contracts-in-test-throwing-contract-is-not-upgrade-safe-use-of-delegatecall-is-not-allowed/32743/5